Moved app state to App component Removed redundant states
